feat(#20): substitute mocked data to real data
- use data returned from api
- remove mocked list
